Upgrade to Laravel 12

+++ src/backend/app/Models/Person/Person.php
@@ -191,7 +191,7 @@ class Person extends BaseModel implements HasAddressesInterface, RoleInterface {
             return false;
         }
 
-        return Carbon::parse($lastPayment->created_at)->diffInDays(now()) <= 25;
+        return (int) Carbon::parse($lastPayment->created_at)->diffInDays(now()) <= 25;
     }
 
     /**
